refactor : 클래스 객체지향적으로 수정
