+++ src/main/java/baseball/Application.java
@@ -1,14 +1,31 @@
 package baseball;
 
 import java.util.*;
+import mallang.missionutils.*;
 
 public class Application {
+
+    static int askFinish() {
+        System.out.println("3개의 숫자를 모두 맞히셨습니다! 게임 종료");
+        System.out.println("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
+        return Integer.parseInt(Console.readLine());
+    }
+
+    static int endOrAgain(int again) {
+        if (again == 1) {
+            return 1;
+        } else if (again == 2) {
+            System.out.println("게임종료");
+            return 0;
+        } else {
+            throw new IllegalArgumentException();
+        }
+    }
+
     public static void main(String[] args) {
 
-        Count count = new Count();
-        HandleUserNum handleUserNum = new HandleUserNum();
-        HandleRandomNum handleRandomNum = new HandleRandomNum();
-        HandleProgress handleProgress = new HandleProgress();
+        Hint hint = new Hint();
+        BaseballNumber baseballNumber = new BaseballNumber();
 
         List<Integer> randomNum;
         List<Integer> userNum;
@@ -18,7 +35,7 @@ public class Application {
 
         while (doGameSet == 1) {
             randomNum = new ArrayList<>();
-            handleRandomNum.pickNum(randomNum);
+            baseballNumber.pickNum(randomNum);
 
             while (doUserSet == 1) {
                 int strike = 0;
@@ -26,19 +43,19 @@ public class Application {
 
                 userNum = new ArrayList<>();
 
-                String[] userArr = handleUserNum.getUserNum();
+                String[] userArr = baseballNumber.getUserNum();
 
-                handleUserNum.checkUserNum(userArr);
-                handleUserNum.addNumToList(userNum, userArr);
+                baseballNumber.checkUserNum(userArr);
+                baseballNumber.addNumToList(userNum, userArr);
 
-                strike = count.countStrike(randomNum, userNum);
-                ball = count.countBall(userNum, randomNum);
+                strike = hint.countStrike(randomNum, userNum);
+                ball = hint.countBall(userNum, randomNum);
 
-                doUserSet = handleProgress.printAnswer(strike, ball);
+                doUserSet = hint.printAnswer(strike, ball);
             }
 
-            int again = handleProgress.askFinish();
-            doGameSet = handleProgress.endOrAgain(again);
+            int again = askFinish();
+            doGameSet = endOrAgain(again);
             doUserSet = 1;
         }
     }
