+++ src/main/java/baseball/Hint.java
@@ -1,9 +1,8 @@
 package baseball;
 
-import java.util.List;
+import java.util.*;
 
-class Count {
-
+public class Hint {
     int countStrike(List<Integer> randomNum, List<Integer> userNum) {
         int cnt = 0;
         for (int i = 0; i < 3; i++) {
@@ -29,5 +28,21 @@ class Count {
             }
         }
         return 0;
+    }
+
+    int printAnswer(int strike, int ball) {
+        if (strike == 3) {
+            System.out.println("3스트라이크");
+            return 0;
+        } else if (strike != 0 && ball != 0) {
+            System.out.printf("%d볼 %d스트라이크\n", ball, strike);
+        } else if (strike != 0 && ball == 0) {
+            System.out.printf("%d스트라이크\n", strike);
+        } else if (strike == 0 && ball != 0) {
+            System.out.printf("%d볼\n", ball);
+        } else {
+            System.out.printf("낫싱\n");
+        }
+        return 1;
     }
 }
